Hotfix attempt two: after checkout, fetch the named git tag corresponding to the checked-out ref
By default and for performance reasons, GitHub actions/checkout@v2 intentionally performs a zero-depth clone of the commit ref provided to it.  For our use case, we would also like to have the git tag corresponding to the release available in the local repository clone so that poetry-dynamic-versioning can discover and use this tag name to determine the version label for the package build.

A command to achieve this was discovered at https://github.com/actions/checkout/issues/217#issuecomment-636481619 (and modified slightly to only fetch the relevant tag)